chore: rename function and interfaces package
