+++ bootstrapper/internal/initserver/BUILD.bazel
@@ -8,7 +8,7 @@ go_library(
     visibility = ["//bootstrapper:__subpackages__"],
     deps = [
         "//bootstrapper/initproto",
-        "//bootstrapper/internal/interfaces",
+        "//bootstrapper/internal/addresses",
         "//bootstrapper/internal/journald",
         "//internal/atls",
         "//internal/attestation",
